refactor: split DetailPanel, Sidebar, StagedModal, and useStaged
Extracts DetailHeader/DetailMetadata, SidebarRow, and the list-diff
components/helpers into their own files, and separates useStaged's
pure staging computations into stagingLogic.ts so they're testable
without React.
